feat(package): Refactor for single module import from package. Added game version collection at instance instantiation. Game data version parameter for get_game_data() now defaults to current version if not supplied.
