+++ .gitignore
@@ -25,7 +25,7 @@ share/python-wheels/
 .installed.cfg
 *.egg
 MANIFEST
-src/utils.py
+src/swgoh_comlink/utils/
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
